Move flatten_signal
